refactor: independ file of python analyse part
