+++ src/build-process/fontforge-py/static/Scripts.txt
